moves all intent launches to respective fragments
